test(content-mapper): stabilize exact LSP oracle
