+++ crates/vize/tests/content_mapper_lsp_support/mod.rs
@@ -2,11 +2,12 @@
 
 use std::{path::Path, str::FromStr};
 
-use corsa_lsp::{LspClient, LspOverlay};
+use corsa_lsp::LspClient;
 use lsp_types::{FileChangeType, FileEvent, Uri};
 use serde_json::{Value, json};
 use vize_carton::String as CompactString;
 
+mod component_oracles;
 mod leak_assertions;
 mod navigation;
 mod package_install;
@@ -14,6 +15,11 @@ mod process_output;
 pub mod raw_requests;
 mod responder;
 mod stop;
+#[allow(unused_imports)]
+pub use component_oracles::{
+    assert_component_completions, assert_component_members, assert_component_navigation,
+    assert_prop_navigation,
+};
 pub use leak_assertions::{assert_no_generated_uri, assert_no_generated_uri_or_zero_range};
 #[allow(unused_imports)]
 pub use navigation::{
@@ -24,6 +30,7 @@ pub use package_install::install_packages;
 #[allow(unused_imports)]
 pub use process_output::output_text;
 pub use responder::EditorResponder;
+#[allow(unused_imports)]
 pub use stop::StopOnDrop;
 
 struct RawDocumentDiagnostic;
@@ -129,104 +136,6 @@ pub async fn assert_completion(
         "{resolved:#}"
     );
     assert!(!resolved_text.contains(".vue.ts"), "{resolved:#}");
-}
-
-pub async fn assert_component_completions(
-    client: &LspClient,
-    overlay: &LspOverlay,
-    uri: &Uri,
-    source: &str,
-) {
-    let partial = source
-        .replace(":count", ":cou")
-        .replace("@save=", "@sa=")
-        .replace("@save-item", "@save-");
-    let prop_position = position(&partial, partial.find(":cou").unwrap() + ":cou".len());
-    let event_position = position(&partial, partial.find("@sa").unwrap() + "@sa".len());
-    let kebab_event_position = position(&partial, partial.find("@save-").unwrap() + "@save-".len());
-    overlay.replace(uri, partial.as_str()).unwrap();
-    let uri_text = uri.as_str();
-    assert_completion(
-        client,
-        uri_text,
-        &prop_position,
-        "count",
-        &["(property) count: number"],
-    )
-    .await;
-    assert_completion(client, uri_text, &event_position, "save", &["number"]).await;
-    assert_completion(
-        client,
-        uri_text,
-        &kebab_event_position,
-        "save-item",
-        &["string"],
-    )
-    .await;
-    overlay.replace(uri, source).unwrap();
-}
-
-pub async fn assert_component_navigation(
-    client: &LspClient,
-    uri: &str,
-    position: &Value,
-    component_name: &str,
-    target_uri: &str,
-) {
-    let component_hover = hover(client, uri, position).await;
-    let hover_text = serde_json::to_string(&component_hover).unwrap();
-    assert!(
-        hover_text.contains(component_name) && !hover_text.contains("__vize_component__"),
-        "{component_hover:#}"
-    );
-    let component_definition = definition(client, uri, position).await;
-    assert_no_generated_uri(&component_definition);
-    let definition_text = serde_json::to_string(&component_definition).unwrap();
-    assert!(
-        definition_text.contains(target_uri),
-        "{component_definition:#}"
-    );
-    assert!(
-        !definition_text.contains(".vue.ts"),
-        "{component_definition:#}"
-    );
-}
-
-pub async fn assert_prop_navigation(
-    client: &LspClient,
-    uri: &str,
-    position: &Value,
-    prop_name: &str,
-    prop_type: &str,
-    target_uri: &str,
-    target_start: &Value,
-) {
-    let prop_hover = hover(client, uri, position).await;
-    let hover_text = serde_json::to_string(&prop_hover).unwrap();
-    assert!(
-        hover_text.contains(prop_name) && hover_text.contains(prop_type),
-        "{prop_hover:#}"
-    );
-    let prop_definition = definition(client, uri, position).await;
-    assert_no_generated_uri_or_zero_range(&prop_definition);
-    assert!(
-        contains_location(&prop_definition, target_uri, target_start),
-        "{prop_definition:#}"
-    );
-}
-pub async fn assert_component_members(
-    client: &LspClient,
-    parent: (&str, &str),
-    child: (&str, &str),
-) {
-    for (usage, declaration, name, ty) in [
-        (":count", "count: number", "count", "number"),
-        ("@save", "save: [value", "save", "number"),
-    ] {
-        let usage = position(parent.1, parent.1.find(usage).unwrap() + 1);
-        let declaration = position(child.1, child.1.find(declaration).unwrap());
-        assert_prop_navigation(client, parent.0, &usage, name, ty, child.0, &declaration).await;
-    }
 }
 
 pub async fn pull_diagnostics(client: &LspClient, uri: &str) -> Value {
